fix(migration): Ensure biodata_siswas table is created only if it does not exist

+++ database/migrations/2025_03_03_035918_create_biodata_siswas_table.php
@@ -11,7 +11,8 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::createIfNotExists('biodata_siswas', function (Blueprint $table) {
+        if (!Schema::hasTable('biodata_siswas')) {
+            Schema::create('biodata_siswas', function (Blueprint $table) {
             $table->id();
             $table->unsignedBigInteger('walas_id');
             $table->foreign('walas_id')->references('id')->on('walas')->onDelete('cascade')->onUpdate ('cascade');
@@ -63,7 +64,8 @@ return new class extends Migration
             $table->text('pengalaman_ekskul');
             $table->text('kepribadian');
             $table->timestamps();
-        });
+            });
+        }
     }
 
     /**
